fix(daemon): own explain-request wire contract in protocol layer to keep transport graph kernel-free
